+++ src/starlite_saqlalchemy/compression.py
@@ -4,4 +4,4 @@ from __future__ import annotations
 from starlite.config.compression import CompressionConfig
 
 config = CompressionConfig(backend="gzip")
-"""Default compression config"""
+"""Default compression config."""