+++ src/starlite_saqlalchemy/exceptions.py
@@ -17,13 +17,6 @@ from starlite.middleware.exceptions.debug_response import create_debug_response
 from starlite.utils.exception import create_exception_response
 from structlog.contextvars import bind_contextvars
 
-from starlite_saqlalchemy.repository.exceptions import (
-    RepositoryConflictException,
-    RepositoryException,
-    RepositoryNotFoundException,
-)
-from starlite_saqlalchemy.service import ServiceException, UnauthorizedException
-
 if TYPE_CHECKING:
     from typing import Any
 
@@ -33,16 +26,38 @@ if TYPE_CHECKING:
     from starlite.types import Scope
     from starlite.utils.exception import ExceptionResponseContent
 
-__all__ = ["after_exception_hook_handler"]
+__all__ = (
+    "AuthorizationError",
+    "ConflictError",
+    "NotFoundError",
+    "StarliteSaqlalchemyError",
+    "after_exception_hook_handler",
+)
 
 
-class ConflictException(HTTPException):
+class StarliteSaqlalchemyError(Exception):
+    """Base exception type for the lib's custom exception types."""
+
+
+class ConflictError(StarliteSaqlalchemyError):
+    """Exception for data integrity errors."""
+
+
+class NotFoundError(StarliteSaqlalchemyError):
+    """Referenced identity doesn't exist."""
+
+
+class AuthorizationError(StarliteSaqlalchemyError):
+    """A user tried to do something they shouldn't have."""
+
+
+class _HTTPConflictException(HTTPException):
     """Request conflict with the current state of the target resource."""
 
     status_code = 409
 
 
-class ForbiddenException(HTTPException):
+class _HTTPForbiddenException(HTTPException):
     """Server understands the request but refuses to authorize it."""
 
     status_code = 403
@@ -64,8 +79,8 @@ async def after_exception_hook_handler(exc: Exception, _scope: Scope, _state: St
     bind_contextvars(exc_info=sys.exc_info())
 
 
-def repository_exception_to_http_response(
-    request: Request[Any, Any], exc: RepositoryException
+def starlite_saqlalchemy_exception_to_http_response(
+    request: Request[Any, Any], exc: StarliteSaqlalchemyError
 ) -> Response[ExceptionResponseContent]:
     """Transform repository exceptions to HTTP exceptions.
 
@@ -77,32 +92,12 @@ def repository_exception_to_http_response(
         Exception response appropriate to the type of original exception.
     """
     http_exc: type[HTTPException]
-    if isinstance(exc, RepositoryNotFoundException):
+    if isinstance(exc, NotFoundError):
         http_exc = NotFoundException
-    elif isinstance(exc, RepositoryConflictException):
-        http_exc = ConflictException
-    else:
-        http_exc = InternalServerException
-    if http_exc is InternalServerException and request.app.debug:
-        return create_debug_response(request, exc)
-    return create_exception_response(http_exc())
-
-
-def service_exception_to_http_response(
-    request: Request[Any, Any], exc: ServiceException
-) -> Response[ExceptionResponseContent]:
-    """Transform service exceptions to HTTP exceptions.
-
-    Args:
-        request: The request that experienced the exception.
-        exc: Exception raised during handling of the request.
-
-    Returns:
-        Exception response appropriate to the type of original exception.
-    """
-    http_exc: type[HTTPException]
-    if isinstance(exc, UnauthorizedException):
-        http_exc = ForbiddenException
+    elif isinstance(exc, ConflictError):
+        http_exc = _HTTPConflictException
+    elif isinstance(exc, AuthorizationError):
+        http_exc = _HTTPForbiddenException
     else:
         http_exc = InternalServerException
     if http_exc is InternalServerException and request.app.debug: