+++ src/starlite_saqlalchemy/init_plugin.py
@@ -35,7 +35,6 @@ from typing import TYPE_CHECKING, Any, TypeVar
 from pydantic import BaseModel
 from starlite.app import DEFAULT_CACHE_CONFIG, DEFAULT_OPENAPI_CONFIG
 from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
-from starlite.response import Response
 from structlog.types import Processor  # noqa: TC002
 
 from starlite_saqlalchemy import (
@@ -52,13 +51,13 @@ from starlite_saqlalchemy import (
     sqlalchemy_plugin,
 )
 from starlite_saqlalchemy.health import health_check
-from starlite_saqlalchemy.repository.exceptions import RepositoryException
-from starlite_saqlalchemy.serializer import default_serializer
-from starlite_saqlalchemy.service import ServiceException, make_service_callback
+from starlite_saqlalchemy.service import make_service_callback
+from starlite_saqlalchemy.type_encoders import type_encoders_map
 from starlite_saqlalchemy.worker import create_worker_instance
 
 if TYPE_CHECKING:
     from starlite.config.app import AppConfig
+    from starlite.types import TypeEncodersMap
 
 T = TypeVar("T")
 
@@ -76,90 +75,90 @@ class PluginConfig(BaseModel):
     worker_functions: list[Callable[..., Any] | tuple[str, Callable[..., Any]]] = [
         (make_service_callback.__qualname__, make_service_callback)
     ]
-    """
-    Queue worker functions.
-    """
+    """Queue worker functions."""
     do_after_exception: bool = True
-    """
+    """Configure after exception handler.
+
     Add the hook handler to
     [`AppConfig.after_exception`][starlite.config.app.AppConfig.after_exception].
     """
     do_cache: bool = True
-    """
+    """Configure redis cache backend.
+
     Add configuration for the redis-backed cache to
     [`AppConfig.cache_config`][starlite.config.app.AppConfig.cache_config].
     """
     do_compression: bool = True
-    """
+    """Confiture compression backend.
+
     Add configuration for gzip compression to
     [`AppConfig.compression_config`][starlite.config.app.AppConfig.compression_config].
     """
     do_collection_dependencies = True
-    """
+    """Add collection route dependencies.
+
     Add the [`Provide`][starlite.datastructures.Provide]'s for collection route dependencies to
     [`AppConfig.dependencies`][starlite.config.app.AppConfig.dependencies].
     """
     do_exception_handlers: bool = True
-    """
+    """Configure exception handlers.
+
     Add the repository/service exception http translation handlers to
     [`AppConfig.exception_handlers`][starlite.config.app.AppConfig.exception_handlers].
     """
     do_health_check: bool = True
-    """
+    """Configure a health check.
+
     Add the health check controller to
     [`AppConfig.route_handlers`][starlite.config.app.AppConfig.route_handlers].
     """
     do_logging: bool = True
-    """
+    """Configure logging.
+
     Set the logging configuration object to
     [`AppConfig.logging_config`][starlite.config.app.AppConfig.logging_config].
     """
     do_openapi: bool = True
-    """
+    """Configure OpenAPI.
+
     Set the OpenAPI config object to
     [`AppConfig.openapi_config`][starlite.config.app.AppConfig.openapi_config].
     """
-    do_response_class: bool = True
-    """
-    Set the custom response class to
-    [`AppConfig.response_class`][starlite.config.app.AppConfig.response_class].
-    """
     do_sentry: bool = True
-    """
+    """Configure sentry.
+
     Configure the application to initialize Sentry on startup. Adds a handler to
     [`AppConfig.on_startup`][starlite.config.app.AppConfig.on_startup].
     """
     do_set_debug: bool = True
-    """
+    """Configure Starlite debug mode.
+
     Allow the plugin to set the starlite `debug` parameter. Parameter set to value of
     [`AppConfig.debug`][starlite_saqlalchemy.settings.AppSettings.DEBUG].
     """
     do_sqlalchemy_plugin: bool = True
-    """
+    """Configure SQLAlchemy plugin.
+
     Set the SQLAlchemy plugin on the application. Adds the plugin to
     [`AppConfig.plugins`][starlite.config.app.AppConfig.plugins].
     """
+    do_type_encoders: bool = True
+    """Configure custom type encoders on the app."""
+
     do_worker: bool = True
-    """
-    Configure the async worker on the application. This action instantiates a worker instance and
-    sets handlers for [`AppConfig.on_startup`][starlite.config.app.AppConfig.on_startup] and
+    """Configure the async worker on the application.
+
+    This action instantiates a worker instance and sets handlers for
+    [`AppConfig.on_startup`][starlite.config.app.AppConfig.on_startup] and
     [`AppConfig.on_shutdown`][starlite.config.app.AppConfig.on_shutdown] that manage the lifecycle
     of the `SAQ` worker.
-    """
-    serializer: Callable[[Any], Any] = default_serializer
-    """
-    The serializer callable that is used by the custom [`Response`][starlite.response.Response]
-    class that is created.
-    If [`AppConfig.response_class`][starlite.config.app.AppConfig.response_class] is not `None`,
-    this is ignored.
-    If
-    [`PluginConfig.do_response_class`][starlite_saqlalchemy.init_plugin.PluginConfig.do_response_class]
-    is `False`, this is ignored.
     """
     # the addition of the health check filter processor makes mypy think `log.default_processors` is
     # list[object].. seems typed correctly to me :/
     log_processors: Sequence[Processor] = log.default_processors  # type:ignore[assignment]
     """Chain of structlog log processors."""
+    type_encoders: TypeEncodersMap = type_encoders_map
+    """Map of type to serializer callable."""
 
 
 class ConfigureApp:
@@ -194,12 +193,12 @@ class ConfigureApp:
         self.configure_health_check(app_config)
         self.configure_logging(app_config)
         self.configure_openapi(app_config)
-        self.configure_response_class(app_config)
         self.configure_sentry(app_config)
         self.configure_sqlalchemy_plugin(app_config)
+        self.configure_type_encoders(app_config)
         self.configure_worker(app_config)
 
-        app_config.on_shutdown.extend([http.Client.close, redis.client.close])
+        app_config.on_shutdown.extend([http.on_shutdown, redis.client.close])
         return app_config
 
     def configure_after_exception(self, app_config: AppConfig) -> None:
@@ -250,7 +249,7 @@ class ConfigureApp:
             app_config.compression_config = compression.config
 
     def configure_debug(self, app_config: AppConfig) -> None:
-        """Set Starlite's `debug` parameter.
+        """Set the Starlite `debug` parameter.
 
         Args:
             app_config: The Starlite application config object.
@@ -271,10 +270,8 @@ class ConfigureApp:
             return
 
         app_config.exception_handlers.setdefault(
-            RepositoryException, exceptions.repository_exception_to_http_response
-        )
-        app_config.exception_handlers.setdefault(
-            ServiceException, exceptions.service_exception_to_http_response
+            exceptions.StarliteSaqlalchemyError,
+            exceptions.starlite_saqlalchemy_exception_to_http_response,
         )
 
     def configure_health_check(self, app_config: AppConfig) -> None:
@@ -310,20 +307,6 @@ class ConfigureApp:
         if self.config.do_openapi and app_config.openapi_config == DEFAULT_OPENAPI_CONFIG:
             app_config.openapi_config = openapi.config
 
-    def configure_response_class(self, app_config: AppConfig) -> None:
-        """Add the custom response class.
-
-        No-op if the [`AppConfig.response_class`][starlite.config.app.AppConfig.response_class]
-        is not `None`.
-
-        Args:
-            app_config: The Starlite application config object.
-        """
-        if self.config.do_response_class and app_config.response_class is None:
-            app_config.response_class = type(
-                "Response", (Response,), {"serializer": staticmethod(self.config.serializer)}
-            )
-
     def configure_sentry(self, app_config: AppConfig) -> None:
         """Add handler to configure Sentry integration.
 
@@ -344,6 +327,15 @@ class ConfigureApp:
         """
         if self.config.do_sqlalchemy_plugin:
             app_config.plugins.append(SQLAlchemyPlugin(config=sqlalchemy_plugin.config))
+
+    def configure_type_encoders(self, app_config: AppConfig) -> None:
+        """Set mapping of type encoders on the application config.
+
+        Args:
+            app_config: The Starlite application config object.
+        """
+        if self.config.do_type_encoders:
+            app_config.type_encoders = self.config.type_encoders
 
     def configure_worker(self, app_config: AppConfig) -> None:
         """Configure the `SAQ` async worker.