+++ src/starlite_saqlalchemy/log/controller.py
@@ -31,6 +31,7 @@ LOGGER = structlog.get_logger()
 
 HTTP_RESPONSE_START: Literal["http.response.start"] = "http.response.start"
 HTTP_RESPONSE_BODY: Literal["http.response.body"] = "http.response.body"
+REQUEST_BODY_FIELD: Literal["body"] = "body"
 
 
 def drop_health_logs(_: WrappedLogger, __: str, event_dict: EventDict) -> EventDict:
@@ -108,8 +109,8 @@ class BeforeSendHandler:
             extract_scheme="scheme" in settings.log.REQUEST_FIELDS,
             obfuscate_cookies=settings.log.OBFUSCATE_COOKIES,
             obfuscate_headers=settings.log.OBFUSCATE_HEADERS,
-            parse_body=True,
-            parse_query=True,
+            parse_body=False,
+            parse_query=False,
         )
         self.response_extractor = ResponseDataExtractor(
             extract_body="body" in settings.log.RESPONSE_FIELDS,
@@ -138,11 +139,16 @@ class BeforeSendHandler:
         # ignore intermediate content of streaming responses for now.
         elif message["type"] == HTTP_RESPONSE_BODY and message["more_body"] is False:
             scope["state"][HTTP_RESPONSE_BODY] = message
-            if self.do_log_request:
-                await self.log_request(scope)
-            if self.do_log_response:
-                await self.log_response(scope)
-            await LOGGER.alog(scope["state"]["log_level"], settings.log.HTTP_EVENT)
+            try:
+                if self.do_log_request:
+                    await self.log_request(scope)
+                if self.do_log_response:
+                    await self.log_response(scope)
+                await LOGGER.alog(scope["state"]["log_level"], settings.log.HTTP_EVENT)
+            except Exception as exc:  # pylint: disable=broad-except
+                # just in-case something in the context causes the error
+                structlog.contextvars.clear_contextvars()
+                await LOGGER.aerror("Error in logging before-send handler!", exc_info=exc)
 
     async def log_request(self, scope: "Scope") -> None:
         """Handle extracting the request data and logging the message.
@@ -185,7 +191,14 @@ class BeforeSendHandler:
             if value is missing:  # pragma: no cover
                 continue
             if isawaitable(value):
-                value = await value
+                # Prevent Starlite from raising a RuntimeError
+                # when trying to read an empty request body.
+                try:
+                    value = await value
+                except RuntimeError:
+                    if key != REQUEST_BODY_FIELD:
+                        raise
+                    value = None
             data[key] = value
         return data